+++ backend/src/app.module.ts
@@ -3,6 +3,10 @@ import { ConfigModule } from '@nestjs/config';
 import { DatabaseModule } from './config/database.module';
 import { AuthModule } from './auth/auth.module';
 import { UsersModule } from './users/users.module';
+import { GroupsModule } from './groups/groups.module';
+import { ReservationsModule } from './reservations/reservations.module';
+import { ProfileModule } from './profile/profile.module';
+import { NotificationsModule } from './notifications/notifications.module';
 
 @Module({
   imports: [
@@ -12,6 +16,10 @@ import { UsersModule } from './users/users.module';
     DatabaseModule,
     AuthModule,
     UsersModule,
+    GroupsModule,
+    ReservationsModule,
+    ProfileModule,
+    NotificationsModule,
   ],
 })
 export class AppModule {}