+++ backend/src/reservations/reservations.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Reservation } from '../models/reservation.entity';
+import { ReservationsController } from './reservations.controller';
+import { ReservationsService } from './reservations.service';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([Reservation])],
+  controllers: [ReservationsController],
+  providers: [ReservationsService],
+})
+export class ReservationsModule {}